auth_api/requirements.txt
click==8.0.1
Flask==2.0.1
importlib-metadata==4.7.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyJWT==2.1.0
PyMySQL==1.0.2
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
